+++ coursework/src/game/player/Player.h
@@ -51,7 +51,7 @@ public:
 
         ManipulationOfPolygon::movePolygon(_element->getPolygon(), direction_vector);
     }
-    void setSpriteIndex(const unsigned int sprite_index) const { _element->setSpriteIndex(sprite_index); }
+    void setSpriteIndex(const size_t sprite_index) const { _element->setSpriteIndex(sprite_index); }
     
     const Types::Control &getControl() const { return _control; }
     Element &getElement() const { return *_element; }
